opc: transplant opc module from python-pptx
Necessarily a big hairy commit, made somewhat worse because it entailed
also converting from objectify to etree and updating most of the
unitutil mocking functions. Most of the files are new, so at least that
makes things a bit simpler.

This brings the packaging logic local, removing the dependency on
python-opc. This is to avoid package syncronization problems, at least
until the packaging logic becomes boringly stable.