oxml: add XmlString for xml __eq__ override
lxml on Python 3 is not so predictable as to how it orders namespace
declarations when serializing XML output. Because the unit tests
generally rely on a string compare to confirm to XML elements are
equal, this override on the OxmlBaseElement.xml.__eq__() is required to
avoid false positives when namespace and/or attribute order differs but
their contents do not.