run: convert CT_Underline.val getter to from_xml()
The new XmlEnumeration base class provides .from_xml() to encapsulate
the mapping of XML attribute values to enumeration values, eliminating
the need to place it in the oxml and repeat it in both directions, one
for each of the getter and the setter.

The predicate in the Run.underline getter unit test needs to be on 'is'
rather than '==' in order to distinguish a `True` value from a 1 value
and likewise for `False` vs. 0.

For now the specialized mappings between WD_UNDERLINE.SINGLE and True
and between WD_UNDERLINE.NONE and False are provided in the oxml
getter and setter directly. Later it might make sense to add a mechanism
to XmlEnumeration to allow adding get and set overrides/aliases, let's
see how it goes and whether that happens often enough to warrant adding
special-case functionality to XmlEnumeration.