txt: refactor Paragraph.insert_paragraph_before()
Extract _insert_paragraph_before() helper method. The helper method can
be tested integrated with oxml. The remaining part of the method is
tested with mocks.

Aside from being good hygiene, this removes the integration dependency
on paragraph.style, allowing that property to be reimplemented without
causing a fail in the _insert_paragraph_before unit tests.