Added font property to paragraph
Paragraphs <w:pPr> can include the same formatting as runs using <w:rPr>
members. This patches the CT_PPr object to map the rPr member and adds a
font member to Paragraph to access it, similar to how it works for runs
and the font member. It is functional as far as I know, but no additions
to the testing modules at this point.