doc: Document.add_section() strips hdr/ftr refs
The implementation of `Document.add_section()` "clones" the previously
last `w:sectPr` element to produce the `w:sectPr` for the new section.
This has the effect of duplicating the header and footer references
along with the rIds used to link them to their header or footer part.

There have been no reports that this causes a problem, but in any case
the correct behavior is to add a new section without any header or
footer definitions, such that the new section "inherits" all
header/footer definitions (and header/footer parts are not "reused").

Add code to remove all `w:headerReference` and `w:footerReference`
elements from a "cloned" `w:sectPr` element when adding a section.