rfctr: extract default .docx template to directory
Maintaining the default document template as a directory allows its
individual parts (XML files) to be curated like code, with line-changes
tracked by Git.

Extract the default template to a directory and load that directory as
the default when no template document is specified to the `Document()`
constructor.

Also add Header and Footer styles.