rfctr: bulk test-layout updates
Back in the day I thought it better to extract a test-specific fixture
for each test. That had the benefit of making parameterized tests fold
cleanly without the parameters in the decorator. The downside was you
had to look in two places to understand the test so I dropped that
practice. Fix some of these I need to change or add to later.