rfctr: resolve StoryChild conflation
We need one type for the base-class role and a different one for the
parameter-type role. The base-class is concrete, the parameter-type is
abstract.

Introduce `docx.types.ProvidesStoryPart` for the parameter-type role.
Add `ProvidesXmlPart` while we're at it for use by `ElementProxy`, which
probably needs a little more thought too.