fix(table): fix _Row.cells can raise IndexError
The original implementation of `_Row.cells` did not take into account
the fact that rows could include unoccupied grid cells at the beginning
and/or end of the row.

This "advanced" feature of tables is sometimes used by the Word table
layout algorithm when the user does not carefully align the right
boundary of cells during resizing, so while quite unusual to be used on
purpose, this arises with some frequency in human-authored documents in
the wild.

The prior implementation of `_Row.cells` assumed that `Table.cells()`
was uniform and the cells for a row could be reliably be computed from
the table column-count and row and column offsets. That assumption
does not always hold and can raise `IndexError` when omitted cells are
present.

This reimplementation remedies that situation. As a side-effect it
should also perform much better when reading large tables.