fix: remove redundant w:pic "insertion"
The removed `._insert_pic(pic)` call is redundant because the `pic`
element is already inserted by the prior `CT_Inline.new()` call. The
reason the second `._insert_pic()` call did not add a second `pic`
element is that the `pic` element is the same instance in both cases. So
that `pic` element is "moved" from where it was, back to the same place.